+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changed:
 Fixed:
 - Initial enabled state for auto-expand buttons
 - Crash on broken symlinks
+- `Cmd+F` search if focus was on meta info
 
 
 ## [1.2.0] – 2026-04-14